% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_accesshistory}
\alias{get_accesshistory}
\title{Get all access list till date}
\usage{
get_accesshistory(pin_board, pin_name)
}
\arguments{
\item{pin_board}{Pin board}

\item{pin_name}{Pin Name}
}
\description{
Get all access list till date
}
