% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_accesslist}
\alias{get_accesslist}
\title{Get access list at a given date or in a specific time period}
\usage{
get_accesslist(pin_board, pin_name, datemin, datemax = NA)
}
\arguments{
\item{pin_board}{Pin board}

\item{pin_name}{pin name}

\item{datemin}{Date in "YYYY-MM-DD" format}

\item{datemax}{Date, either NA or in "YYYY-MM-DD" format}
}
\value{
A list of dataframes with access details
}
\description{
Get access list at a given date or in a specific time period
}
