% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_admins}
\alias{get_admins}
\title{Gets the list of admins}
\usage{
get_admins(pin_board, pin_name, admin_panel)
}
\arguments{
\item{pin_board}{Pin board}

\item{pin_name}{pin name}

\item{admin_panel}{Admin panel name}
}
\description{
Gets the list of admins
}
