% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_board}
\alias{get_board}
\title{Gather pin_board elements}
\usage{
get_board(pin_board, pin_name)
}
\arguments{
\item{pin_board}{pin board}

\item{pin_name}{pin name}
}
\value{
A list with access_panels, access_units, access_df, access_list
}
\description{
Gather pin_board elements
}
