% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_granted_units}
\alias{get_granted_units}
\title{Helper function to get access units/panels with access for a user from an
existing pin board}
\usage{
get_granted_units(user_id, pin_board, pin_name)
}
\arguments{
\item{user_id}{user id}

\item{pin_board}{pin board}

\item{pin_name}{pin name}
}
\value{
A vector with access units
}
\description{
Helper function to get access units/panels with access for a user from an
existing pin board
}
