% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_user_api}
\alias{get_user_api}
\title{Get user data using API}
\usage{
get_user_api(contact_info, url, api_key = NULL)
}
\arguments{
\item{contact_info}{User entered search text}

\item{url}{URL}

\item{api_key}{Valid api key or NULL}
}
\value{
A tibble with user id and username
}
\description{
Get user data using API
}
