% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_iam.R
\name{module_iam_server}
\alias{module_iam_server}
\title{Server logic of module_iam}
\usage{
module_iam_server(id, rAccess_obj)
}
\arguments{
\item{id}{Module's ID}

\item{rAccess_obj}{New instance of rAccess(R6 object)}
}
\description{
Server logic of module_iam
}
