% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sub_iam.R
\name{module_sub_iam_server}
\alias{module_sub_iam_server}
\title{Server logic of module_sub_iam}
\usage{
module_sub_iam_server(id, access_panel_id, rAccess_obj)
}
\arguments{
\item{id}{Module's ID}

\item{access_panel_id}{Access panel ID}

\item{rAccess_obj}{New instance of rAccess(R6 object)}
}
\description{
Server logic of module_sub_iam
}
