% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sub_iam.R
\name{module_sub_iam_ui}
\alias{module_sub_iam_ui}
\title{User-interface definition of module_sub_iam}
\usage{
module_sub_iam_ui(id)
}
\arguments{
\item{id}{Module's ID}
}
\description{
This is a Shiny module used by the main shiny web application
}
