% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rconnect_pin_board}
\alias{rconnect_pin_board}
\title{Function to create rconnect pinboard}
\usage{
rconnect_pin_board(server, key)
}
\arguments{
\item{server}{rconnect server}

\item{key}{API KEY to connect with rconnect servers}
}
\value{
a pin_board
}
\description{
Function to create rconnect pinboard
}
