% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{s3_config}
\alias{s3_config}
\title{Set environmental variables to connect to AWS S3 bucket}
\usage{
s3_config(access_key, secret_key, region = "us-east-1")
}
\arguments{
\item{access_key}{Character. AWS access key.}

\item{secret_key}{Character. AWS secret key.}

\item{region}{Character. AWS region. Defaults to \code{"us-east-1"}.}
}
\description{
Set environmental variables to connect to AWS S3 bucket
}
