% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{s3_pinboard}
\alias{s3_pinboard}
\title{Function to create s3 pin_board}
\usage{
s3_pinboard(
  s3_bucket,
  s3_access_key,
  s3_secret_key,
  s3_region = "us-east-1",
  s3_prefix
)
}
\arguments{
\item{s3_bucket}{s3 bucket name}

\item{s3_access_key}{access key to connect to s3 bucket}

\item{s3_secret_key}{secret key to connect to s3 bucket}

\item{s3_region}{s3 bucket region}

\item{s3_prefix}{Prefix to}
}
\value{
s3 pin board
}
\description{
Function to create s3 pin_board
}
