% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{use_config}
\alias{use_config}
\title{Function to add rAccess configuration file to the given directory}
\usage{
use_config(file_name = "rAccess.yml", path = getwd())
}
\arguments{
\item{file_name}{Config file name with .yml extension}

\item{path}{Directory to which config file is to be added}
}
\description{
Copies a configuration file (.yml) from the package's config directory to a
specified path.
}
